no extra files or comments + refreshCall needs to be a function or error is thrown
