+++ src/index.ts
@@ -1,13 +1,22 @@
 import {AxiosInstance, AxiosResponse} from "axios";
-import {AxiosAuthRefreshOptions, AxiosAuthRefreshCache} from "./types";
-import {createRefreshCall, createRequestQueueInterceptor, mergeConfigs, shouldInterceptError} from "./helpers";
 
-/** @type {AxiosAuthRefreshOptions} */
+// Types
+
+export interface AxiosAuthRefreshOptions {
+    statusCodes?: Array<number>
+}
+
+export interface AxiosAuthRefreshCache {
+    refreshCall: Promise<any>|undefined,
+    requestQueueInterceptorId: number|undefined
+}
+
+// Constants
+
 const defaults: AxiosAuthRefreshOptions = {
     statusCodes: [ 401 ]
 };
 
-/** @type {AxiosAuthRefreshCache} */
 const cache: AxiosAuthRefreshCache = {
     refreshCall: undefined,
     requestQueueInterceptorId: undefined
@@ -23,17 +32,16 @@ const cache: AxiosAuthRefreshCache = {
  * @param {AxiosAuthRefreshOptions} options - options for the interceptor @see defaultOptions
  * @return {AxiosInstance}
  */
-function createAuthRefreshInterceptor(
+export default function createAuthRefreshInterceptor(
     axios: AxiosInstance,
     refreshTokenCall: (error: any) => Promise<any>,
     options: AxiosAuthRefreshOptions = {}
 ) {
-    const id = axios.interceptors.response.use((res: AxiosResponse) => res, (error: any) => {
+    if (typeof refreshTokenCall !== 'function') {
+        throw new Error('axios-auth-refresh requires `refreshTokenCall` to be a function that returns a promise.');
+    }
 
-        if (typeof refreshTokenCall !== 'function') {
-            console.warn('axios-auth-refresh requires `refreshTokenCall` to be a function that returns a promise.');
-            return Promise.reject(error);
-        }
+    const id = axios.interceptors.response.use((res: AxiosResponse) => res, (error: any) => {
 
         // Rewrite default config
         options = mergeConfigs(options, defaults);
@@ -68,4 +76,48 @@ function createAuthRefreshInterceptor(
     return axios;
 }
 
-export default createAuthRefreshInterceptor;
+/**
+ * Merges two config objects (master rewrites slave)
+ */
+export function mergeConfigs(master: AxiosAuthRefreshOptions, def: AxiosAuthRefreshOptions): AxiosAuthRefreshOptions {
+    return { ...def, ...master };
+}
+
+/**
+ * Returns TRUE: when error.response.status is contained in options.statusCodes
+ * Returns FALSE: when error or error.response doesn't exist or options.statusCodes doesn't include response status
+ */
+export function shouldInterceptError(error: any, options: AxiosAuthRefreshOptions): boolean {
+    return error && error.response && options.statusCodes.includes(+error.response.status);
+}
+
+/**
+ * Creates refresh call if it does not exist or returns the existing one
+ */
+export function createRefreshCall(
+    error: any,
+    fn: (error: any) => Promise<any>,
+    cache: AxiosAuthRefreshCache
+): Promise<any> {
+    if (!cache.refreshCall) {
+        cache.refreshCall = fn(error);
+        if (typeof cache.refreshCall.then !== 'function') {
+            console.warn('axios-auth-refresh requires `refreshTokenCall` to return a promise.');
+            return Promise.reject();
+        }
+    }
+    return cache.refreshCall;
+}
+
+/**
+ * Creates refresh call if it does not exist or returns the existing one
+ */
+export function createRequestQueueInterceptor(axios: AxiosInstance, cache: AxiosAuthRefreshCache): number {
+    if (typeof cache.requestQueueInterceptorId === 'undefined') {
+        cache.requestQueueInterceptorId = axios
+            .interceptors
+            .request
+            .use((request) => cache.refreshCall.then(() => request));
+    }
+    return cache.requestQueueInterceptorId;
+}
