--- src/types.ts
@@ -1,8 +1,0 @@
-export interface AxiosAuthRefreshOptions {
-    statusCodes?: Array<number>
-}
-
-export interface AxiosAuthRefreshCache {
-    refreshCall: Promise<any>|undefined,
-    requestQueueInterceptorId: number|undefined
-}
